+++ packages/accounts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bloque/sdk-accounts
 
+## 0.13.1
+
+### Patch Changes
+
+- @bloque/sdk-core@0.13.1
+
 ## 0.13.0
 
 ### Minor Changes